Fixed circular imports and restructured
