+++ dir_acyc_graph.py
@@ -1,3 +1,4 @@
+from dag_structures import Node, PointNode, SegNode, Leaf
 from dag_case1 import handle_case1
 from dag_case2 import handle_case2
 from dag_case3 import handle_case3
@@ -58,47 +59,16 @@ class DAG:
                     .union({self.find_trapezoids(curr_node.left, seg)})
 
         if isinstance(curr_node, SegNode):
-            if seg.is_above(curr_node.seg):
+            seg_lowest_point = seg.get_lower_point()
+            seg_higher_point = seg.get_higher_point()
+
+            if not seg_higher_point.is_above(curr_node.seg.get_lower_point()):
+                return {self.find_trapezoids(curr_node.right, seg)}
+            elif seg_lowest_point.is_above(curr_node.seg.get_higher_point()):
                 return {self.find_trapezoids(curr_node.left, seg)}
             else:
-                return {self.find_trapezoids(curr_node.right, seg)}
+                return {self.find_trapezoids(curr_node.right, seg)} \
+                    .union({self.find_trapezoids(curr_node.left, seg)})
 
     def create_output_matrix(self):
         return None
-
-
-class Node:
-    def __init__(self, data):
-        self.data = data
-
-
-class Internal:
-    def __init__(self, left, right):
-        self.left = left
-        self.right = right
-
-    def is_child_left(self, node):
-        if self.left == node:
-            return True
-        elif self.right == node:
-            return False
-        else:
-            raise Exception("Not the parent of the child")
-
-
-class PointNode(Internal):
-    def __init__(self, left, right, point):
-        super().__init__(left, right)
-        self.point = point
-
-
-class SegNode(Internal):
-    def __init__(self, left, right, seg):
-        super().__init__(left, right)
-        self.seg = seg
-
-
-class Leaf:
-    def __init__(self, trap):
-        super().__init__()
-        self.trap = trap
