+++ rand_incr_alg.py
@@ -1,11 +1,11 @@
 from random import shuffle
-from trapezoidal_map import Segment, Point, TrapezoidalMap
-
+from dir_acyc_graph import DAG
 
 def randomized_incremental_algorithm(trap_map):
     shuffle(trap_map.line_segments)
     segments = trap_map.line_segments
     num_segments = len(segments)
+    dag = DAG(trap_map.bbox)
     for _ in range(0, num_segments):
         segment = segments.pop(0)
-        trap_map.dag.add_new_segment(segment)
+        dag.add_new_segment(segment)
